set a min-height to the gallery.
Did this so images of different resolutions won't move the footer
around. I also set a top margin so the images aren't up against
the header. The previous and next buttons now also have a black
background since the clear background made it difficult to find
the next and previous buttons images with a light background.